+++ modules/jpa/src/test/kotlin/com/explyt/jpa/ql/reference/kotlin/JpqlExternalReferenceCompletionTest.kt
@@ -70,8 +70,8 @@ abstract class JpqlExternalReferenceCompletionTest : ExplytKotlinLightTestCase()
 
         val name = getTestName(true)
         var vf = myFixture.copyFileToProject(
-            "$name.jpql",
-            "$name.jpql"
+            "$name.ejpql",
+            "$name.ejpql"
         )
         myFixture.configureFromExistingVirtualFile(vf)